# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Preeku Paper Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Wallet {
  id: number;
  balance: number;
  initialBalance: number;
  totalInvested: number;
  totalPnl: number;
  totalPnlPercent: number;
  updatedAt: string;
}

export interface Stock {
  symbol: string;
  name: string;
  exchange: string;
  sector: string;
  currentPrice: number;
  previousClose: number;
  change: number;
  changePercent: number;
  high: number;
  low: number;
  volume: number;
  marketCap: number;
}

export interface PricePoint {
  time: string;
  price: number;
  volume: number;
}

export type StockDetail = Stock & {
  priceHistory: PricePoint[];
};

export interface WatchlistItem {
  id: number;
  symbol: string;
  name: string;
  currentPrice: number;
  change: number;
  changePercent: number;
  addedAt: string;
}

export interface AddWatchlistBody {
  symbol: string;
}

export type OrderOrderType =
  (typeof OrderOrderType)[keyof typeof OrderOrderType];

export const OrderOrderType = {
  MARKET: "MARKET",
  LIMIT: "LIMIT",
} as const;

export type OrderSide = (typeof OrderSide)[keyof typeof OrderSide];

export const OrderSide = {
  BUY: "BUY",
  SELL: "SELL",
} as const;

export type OrderProductType =
  (typeof OrderProductType)[keyof typeof OrderProductType];

export const OrderProductType = {
  INTRADAY: "INTRADAY",
  DELIVERY: "DELIVERY",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  PENDING: "PENDING",
  EXECUTED: "EXECUTED",
  CANCELLED: "CANCELLED",
} as const;

export interface Order {
  id: number;
  symbol: string;
  stockName: string;
  orderType: OrderOrderType;
  side: OrderSide;
  productType: OrderProductType;
  quantity: number;
  price: number;
  executedPrice?: number | null;
  status: OrderStatus;
  totalValue: number;
  pnl?: number | null;
  placedAt: string;
  executedAt?: string | null;
}

export type PlaceOrderBodyOrderType =
  (typeof PlaceOrderBodyOrderType)[keyof typeof PlaceOrderBodyOrderType];

export const PlaceOrderBodyOrderType = {
  MARKET: "MARKET",
  LIMIT: "LIMIT",
} as const;

export type PlaceOrderBodySide =
  (typeof PlaceOrderBodySide)[keyof typeof PlaceOrderBodySide];

export const PlaceOrderBodySide = {
  BUY: "BUY",
  SELL: "SELL",
} as const;

export type PlaceOrderBodyProductType =
  (typeof PlaceOrderBodyProductType)[keyof typeof PlaceOrderBodyProductType];

export const PlaceOrderBodyProductType = {
  INTRADAY: "INTRADAY",
  DELIVERY: "DELIVERY",
} as const;

export interface PlaceOrderBody {
  symbol: string;
  orderType: PlaceOrderBodyOrderType;
  side: PlaceOrderBodySide;
  productType: PlaceOrderBodyProductType;
  quantity: number;
  limitPrice?: number | null;
}

export type PositionProductType =
  (typeof PositionProductType)[keyof typeof PositionProductType];

export const PositionProductType = {
  INTRADAY: "INTRADAY",
  DELIVERY: "DELIVERY",
} as const;

export interface Position {
  id: number;
  symbol: string;
  stockName: string;
  quantity: number;
  avgBuyPrice: number;
  currentPrice: number;
  pnl: number;
  pnlPercent: number;
  currentValue: number;
  investedValue: number;
  productType: PositionProductType;
}

export interface Holding {
  id: number;
  symbol: string;
  stockName: string;
  quantity: number;
  avgBuyPrice: number;
  currentPrice: number;
  pnl: number;
  pnlPercent: number;
  currentValue: number;
  investedValue: number;
  dayChange: number;
  dayChangePercent: number;
}

export interface PerformerStock {
  symbol: string;
  name: string;
  pnlPercent: number;
}

export interface PortfolioSummary {
  totalInvested: number;
  currentValue: number;
  totalPnl: number;
  totalPnlPercent: number;
  dayPnl: number;
  dayPnlPercent: number;
  walletBalance: number;
  openPositions: number;
  totalHoldings: number;
  totalOrders: number;
  bestPerformer?: PerformerStock | null;
  worstPerformer?: PerformerStock | null;
}

export interface HeatmapStock {
  symbol: string;
  name: string;
  changePercent: number;
  marketCap: number;
}

export interface HeatmapSector {
  sector: string;
  changePercent: number;
  stocks: HeatmapStock[];
}

export interface SuccessResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListStocksParams = {
  search?: string;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  EXECUTED: "EXECUTED",
  CANCELLED: "CANCELLED",
  PENDING: "PENDING",
} as const;