# lib/api-zod/src/generated/types/heatmapStock.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Preeku Paper Trading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface HeatmapStock {
  symbol: string;
  name: string;
  changePercent: number;
  marketCap: number;
}