# lib/api-zod/src/generated/types/holding.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Preeku Paper Trading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Holding {
  id: number;
  symbol: string;
  stockName: string;
  quantity: number;
  avgBuyPrice: number;
  currentPrice: number;
  pnl: number;
  pnlPercent: number;
  currentValue: number;
  investedValue: number;
  dayChange: number;
  dayChangePercent: number;
}