# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Preeku Paper Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderOrderType } from "./orderOrderType";
import type { OrderProductType } from "./orderProductType";
import type { OrderSide } from "./orderSide";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  symbol: string;
  stockName: string;
  orderType: OrderOrderType;
  side: OrderSide;
  productType: OrderProductType;
  quantity: number;
  price: number;
  executedPrice?: number | null;
  status: OrderStatus;
  totalValue: number;
  pnl?: number | null;
  placedAt: Date;
  executedAt?: Date | null;
}