# lib/api-zod/src/generated/types/position.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Preeku Paper Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PositionProductType } from "./positionProductType";

export interface Position {
  id: number;
  symbol: string;
  stockName: string;
  quantity: number;
  avgBuyPrice: number;
  currentPrice: number;
  pnl: number;
  pnlPercent: number;
  currentValue: number;
  investedValue: number;
  productType: PositionProductType;
}