# lib/api-zod/src/generated/types/pricePoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Preeku Paper Trading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PricePoint {
  time: string;
  price: number;
  volume: number;
}