# lib/api-zod/src/generated/types/stock.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Preeku Paper Trading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Stock {
  symbol: string;
  name: string;
  exchange: string;
  sector: string;
  currentPrice: number;
  previousClose: number;
  change: number;
  changePercent: number;
  high: number;
  low: number;
  volume: number;
  marketCap: number;
}